refactor the all codes andd widgets folder and separate the theme
