+++ lib/view/widgets/info_chip.dart
@@ -1,0 +1,37 @@
+import 'package:flutter/material.dart';
+
+class InfoChip extends StatelessWidget {
+  final IconData icon;
+  final String label;
+  final String value;
+  final Color bg;
+
+  const InfoChip({
+    super.key,
+    required this.icon,
+    required this.label,
+    required this.value,
+    required this.bg,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      padding: const EdgeInsets.symmetric(horizontal: 12, vertical: 10),
+      decoration: BoxDecoration(
+        color: bg,
+        borderRadius: BorderRadius.circular(12),
+      ),
+      child: Row(
+        mainAxisSize: MainAxisSize.min,
+        children: [
+          Icon(icon, size: 18),
+          const SizedBox(width: 8),
+          Text(label, style: const TextStyle(fontWeight: FontWeight.w600)),
+          const SizedBox(width: 6),
+          Text(value),
+        ],
+      ),
+    );
+  }
+}
