+++ lib/view/widgets/section_header.dart
@@ -1,0 +1,25 @@
+import 'package:flutter/material.dart';
+
+class SectionHeader extends StatelessWidget {
+  final String title;
+  final IconData icon;
+  const SectionHeader(
+    this.title, {
+    super.key,
+    this.icon = Icons.assessment_outlined,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return Row(
+      children: [
+        Icon(icon, size: 18),
+        const SizedBox(width: 6),
+        Text(
+          title,
+          style: const TextStyle(fontWeight: FontWeight.w700, fontSize: 15),
+        ),
+      ],
+    );
+  }
+}
